% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.mesh.colors.R
\name{array.mean}
\alias{array.mean}
\title{Internal function for calculating a mean of data in array format}
\usage{
array.mean(data)
}
\arguments{
\item{data}{an array from which to calculate a mean}
}
\value{
a matrix of N_landmarks x N_dims
}
\description{
Internal function for calculating a mean of data in array format
}
