% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array2row.R
\name{array2row3d}
\alias{array2row3d}
\title{Converts arrays to row formats for 3D data}
\usage{
array2row3d(data)
}
\arguments{
\item{data}{an array of N_landmarks x 3 x N_specimens}
}
\value{

}
\description{
Converts arrays to row formats for 3D data
}
