% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.delaunay.R
\name{flip.delaunay}
\alias{flip.delaunay}
\title{flip delaunay triangulation along the Y axis to match an image}
\usage{
flip.delaunay(delaunay.map, example.image)
}
\arguments{
\item{delaunay.map}{delaunay triangulation object}

\item{example.image}{image that corresponds to the delaunay map}
}
\value{
The function will return delaunay objects with the Y-axis values flipped to deal with non-matching origins between images and landmarks
}
\description{
flip delaunay triangulation along the Y axis to match an image
}
