% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.lms.R
\name{flip.lms}
\alias{flip.lms}
\title{Flips landmarks to the same orientation if you have a mixture of left and right facing data}
\usage{
flip.lms(landmarks, imagedir, side_data, side_code = "right")
}
\arguments{
\item{landmarks}{A landmark array with dimensions N_landmarks x 2 x N_observations}

\item{imagedir}{The path to the working folder of images}

\item{side_data}{An index of which side the landmarks are facing}

\item{side_code}{What you call the side you want to flip (for example, "right" or "R")}
}
\value{
An array of landmark info with dimensions N_landmarks x 2 x N_observations with all data facing the same direction
}
\description{
Flips landmarks to the same orientation if you have a mixture of left and right facing data
}
\examples{
guppy.lms <- tps2array(system.file("extdata", "original_lms.TPS", package = "Colormesh"))
flip.test <- flip.lms(landmarks = guppy.lms, imagedir = paste0(path.package("Colormesh"),"/inst/extdata/cropped_images/"), side_data = c("right", rep("left", 9)))

#plot the flipped example
plot(flip.test[,,1])
points(guppy.lms[,,1])
}
