% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmark.images.R
\name{landmark.images}
\alias{landmark.images}
\title{Read in and convert images to jpeg for digitization in geomorph}
\usage{
landmark.images(
  imagedir,
  image.names,
  nlandmarks,
  writedir = NULL,
  dump.tmp.images = T
)
}
\arguments{
\item{imagedir}{directory of images to measure. Only images with landmarks will be processed. The landmark file names are assumed to exactly match the image names.}

\item{image.names}{A vector of image names to look for in imagedir. These images should be unwarped or deformed to a common reference shape.}

\item{nlandmarks}{number of landmarks to acquire per image.}

\item{writedir}{directory to save converted images. If left NULL, images will be save to a temporary directory}

\item{dump.tmp.images}{Logical value to keep temporarily created jpegs. By default they will be deleted when landmarking is completed.}
}
\value{
This function reads in Raw image formats and converts them to jpeg. The function returns the paths to the images that were converted. These jpeg locations can then be used by geomorph for landmarking within R.
}
\description{
Read in and convert images to jpeg for digitization in geomorph
}
\details{
This function is a wrapper around geomorph::digitize2d() that a) allows for more general image format reading and b) automatically reads in the landmarks into R after saving the TPS files. Geomorph only supports reading in jpeg files, so this function converts all images to jpeg in a temporary directory, calls digitize2d for landmarking and then removes the temporary jpegs upon completion.
}
\examples{
#example place 3 landmarks on one image
ex.landmark <- landmark.images(imagedir = paste0(path.package("Colormesh"),"/extdata/cropped_images/"), image.names = "GPHP_001.tiff", nlandmarks = 3)

}
\seealso{
\code{\link[geomorph]{digitize2d}} (used for landmarking)
}
