% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearize.colors.R
\name{linearize.colors}
\alias{linearize.colors}
\title{Linear color data with approximation from Caves & Johnsen 2017}
\usage{
linearize.colors(rgb.object)
}
\arguments{
\item{rgb.object}{A .TPS file}
}
\value{
An array of sampled colors in linearized space
}
\description{
Linear color data with approximation from Caves & Johnsen 2017
}
