% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.mesh.colors.R
\name{mean.mesh.colors}
\alias{mean.mesh.colors}
\title{Calculating means/group means for color sampled data}
\usage{
\method{mean}{mesh.colors}(mesh.colors.object, covariate_data = NULL, group_labels = NULL)
}
\arguments{
\item{mesh.colors.object}{a mesh.colors object. The result of tps.unwarp when color.sampling was specified}

\item{covariate_data}{dataframe with variables to specify a group mean to calculate}

\item{group_labels}{a vector of factors to group by. Up to two separate factors can be accepted}
}
\value{
a mesh.colors object, with $sampled.colors now being the specified mean
}
\description{
Calculating means/group means for color sampled data
}
