% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot.mesh.colors}
\alias{plot.mesh.colors}
\title{plotting individual specimens after color sampling}
\usage{
\method{plot}{mesh.colors}(mesh.colors.object, individual = 1, visualization_type = "sampled")
}
\arguments{
\item{individual}{which individual from your landmark dataframe you'd like to plot}

\item{visualization_type}{plot raw "sampled" color or "calibrated" color? Sampled is the default.}

\item{x}{an object of class "tri.surf.points". If using this function after color sampling, it will be object$delaunay}

\item{...}{Additional plotting parameters to be passed to plot.default}
}
\value{
A list of class tri.surf.points. $interior is the position of internal (non-perimeter) points generated from triangulation. $perimeter is the initial points submitted for triangulation. $centroids is the final set of centroids from the triangulation. $final.mesh is the last round of triangulation. $point.map is the point map used to give the order of perimeter landmarks.
}
\description{
plotting individual specimens after color sampling
}
