% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot}
\alias{plot}
\title{plot Colormesh data}
\usage{
plot(x, ...)
}
\arguments{
\item{x}{object of class tri.surf.points or mesh.colors}

\item{\dots}{for render.meshDist: additional arguments passed to
\code{\link{par}}.}
}
\description{
plot Colormesh data
}
\seealso{
\code{\link{plot.tri.surf.points}}, \code{\link{plot.mesh.colors}}
}
\author{
J. David Aponte
}
