% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot.tri.surf.points}
\alias{plot.tri.surf.points}
\title{plotting method for objects of class "tri.surf.points"}
\usage{
\method{plot}{tri.surf.points}(
  x,
  style = "points",
  corresponding.image,
  wireframe.color = "black",
  point.color = "red",
  ...
)
}
\arguments{
\item{x}{an object of class "tri.surf.points"}

\item{style}{the type of plot to generate. There are currently 2 options, "points" and "triangulation". "points" is the default and just plots the interior and perimeter points. "triangulation" plots the delaunay triangulation wireframe with centroids highlighted.}

\item{...}{Additional plotting parameters to be passed to plot.default}
}
\value{
A list of class tri.surf.points. $interior is the position of internal (non-perimeter) points generated from triangulation. $perimeter is the initial points submitted for triangulation. $centroids is the final set of centroids from the triangulation. $final.mesh is the last round of triangulation. $point.map is the point map used to give the order of perimeter landmarks.
}
\description{
plotting method for objects of class "tri.surf.points"
}
