% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_plotter.R
\name{point.overlap}
\alias{point.overlap}
\title{Diagnostic tool for looking at point sampling density}
\usage{
point.overlap(delaunay.map, px.radius)
}
\arguments{
\item{delaunay.map}{delaunay triangulation object}

\item{px.radius}{The size of the circular neighborhood (in pixels) to sample color around each triangulated point.}
}
\value{
An index of points pairs that will have overlapping (and redundant) pixel information. Only for the interior points of a delaunay triangulation (not the perimeter)
}
\description{
Diagnostic tool for looking at point sampling density
}
