% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2jpeg.R
\name{raw2jpeg}
\alias{raw2jpeg}
\title{Read in and convert raw images to jpeg for digitization in geomorph}
\usage{
raw2jpeg(imagedir, image.names, writedir)
}
\arguments{
\item{imagedir}{directory of images to measure. Only images with landmarks will be processed. The landmark file names are assumed to exactly match the image names.}

\item{image.names}{A vector of image names to look for in imagedir. These images should be unwarped or deformed to a common reference shape.}

\item{writedir}{directory to save converted images.}
}
\value{
This function reads in Raw image formats and converts them to jpeg. These jpegs can then be used by geomorph for landmarking within R.
}
\description{
Read in and convert raw images to jpeg for digitization in geomorph
}
\examples{
#load an image and convert

#use newly created jpegs for landmarking

#remove tmpdir

}
