% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.tps.R
\name{read.tps}
\alias{read.tps}
\title{read TPS files and convert data to an array}
\usage{
read.tps(data)
}
\arguments{
\item{data}{A .TPS file}
}
\value{
A matrix of the landmarks for each observation
}
\description{
read TPS files and convert data to an array
}
