% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb.calibrate.R
\name{rgb.calibrate}
\alias{rgb.calibrate}
\title{Calibration of existing color sampled data}
\usage{
rgb.calibrate(
  sampled.array,
  imagedir,
  image.names,
  calib.file,
  color.standard.values = NULL,
  px.radius = 2,
  flip.y.values = F
)
}
\arguments{
\item{sampled.array}{Previously sampled data from the rgb.measure function.}

\item{imagedir}{directory of images to measure for calibration. Only images with landmarks will be processed. The landmark file names are assumed to exactly match the image names.}

\item{image.names}{A vector of image names to look for in imagedir.}

\item{color.standard.values}{a matrix of known values for collected color standard points. Should be N_points x 3}

\item{px.radius}{The size of the circular neighborhood (in pixels) to sample color around each triangulated point.}

\item{flip.y.values}{should the calbration points be flipped to match the images?}
}
\value{
The function will return $sampled.color-- an N_points x 3 (RGB) x N_observations array of sampled color values. A tri.surf.points class object will also be returned as $delaunay. Finally, a calibrated array of color values will be returned under $calibrated
}
\description{
Calibration of existing color sampled data
}
