% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb.measure.R
\name{rgb.measure}
\alias{rgb.measure}
\title{Color sampling from a set of pre-warped images}
\usage{
rgb.measure(imagedir, image.names, delaunay.map, px.radius = 2)
}
\arguments{
\item{imagedir}{directory of images to measure. Only images with landmarks will be processed. The landmark file names are assumed to exactly match the image names.}

\item{image.names}{A vector of image names to look for in imagedir. These images should be unwarped or deformed to a common reference shape.}

\item{delaunay.map}{delaunay triangulation object}

\item{px.radius}{The size of the circular neighborhood (in pixels) to sample color around each triangulated point.}
}
\value{
If write.images is true, warped images will be saved to the write.dir directory. If color sampling is true, the function will return $sampled.color-- an N_points x 3 (RGB) x N_observations array of sampled color values. A tri.surf.points class object will also be returned as $delaunay.
}
\description{
Color sampling from a set of pre-warped images
}
