% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.circle.R
\name{sampling.circle}
\alias{sampling.circle}
\title{Internal function for color sampling an image. Calculates the pixels in a circular neighborhood of specified radius.}
\usage{
sampling.circle(px.radius = 2)
}
\arguments{
\item{px.radius}{How many neighboring pixels to include when analyzing image color at a given landmark. A value of 1 just gives back the original pixel.}
}
\value{
The coordinates of a circle with radius px.radius
}
\description{
Internal function for color sampling an image. Calculates the pixels in a circular neighborhood of specified radius.
}
