% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tps2array.R
\name{tps2array}
\alias{tps2array}
\title{convert TPS data to landmark array}
\usage{
tps2array(tpsfile)
}
\arguments{
\item{tpsfile}{A read.tps() object}
}
\value{
An array of landmark info with dimensions N_landmarks x 2 x N_observations
}
\description{
convert TPS data to landmark array
}
