% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.plot.R
\name{calib.plot}
\alias{calib.plot}
\title{Diagnostic tool for looking at calibration images and the sampling landmarks}
\usage{
calib.plot(
  imagedir,
  image.names,
  calib.file,
  cex = 2,
  col = "red",
  individual = 1,
  flip.y.values = F
)
}
\arguments{
\item{imagedir}{path to calibration images}

\item{image.names}{A vector of image names to look for in imagedir.}

\item{calib.file}{tps2array object with the sampled color standard. Make sure the IMAGE data in the tps file matches the names of the images exactly (eg. get rid of any paths before the name).}

\item{cex}{change the point size on the image}

\item{col}{change point colors}

\item{individual}{which specimen do you want to plot? Defaults to the first specimen}

\item{flip.y.values}{logical value to try flipping the Y values of the coordinates to see if the landmarks will match up with the color standard}
}
\value{
a plot to check if you should flip the y-axis of the color standard landmarks
}
\description{
Diagnostic tool for looking at calibration images and the sampling landmarks
}
\examples{
#load covariates and calibration file
specimen.factors <- read.csv(system.file("extdata", "specimen_factors.csv", package = "Colormesh"), header = F)
calib.file <- tps2array(system.file("extdata", "calib_LM_coords.TPS", package = "Colormesh"))

calib.plot(paste0(path.package("Colormesh"),"/extdata/original_images/"), image.names = specimen.factors[,1], calib.file = calib.file)

}
