% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reader.R
\name{image_reader}
\alias{image_reader}
\title{Internal tool to read in various image formats for the Colormesh pipeline}
\usage{
image_reader(imagedir, image.names)
}
\arguments{
\item{imagedir}{directory of images.}

\item{image.names}{A vector of image names to look for in imagedir.}
}
\value{
Returns the loaded image
}
\description{
Internal tool to read in various image formats for the Colormesh pipeline
}
