% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.colormesh.dataset.R
\name{make.colormesh.dataset}
\alias{make.colormesh.dataset}
\title{Combine metadata with calibrated colormesh data for analysis}
\usage{
make.colormesh.dataset(
  df,
  specimen.factors,
  use.perimeter.data = F,
  write2csv = NULL
)
}
\arguments{
\item{df}{A calibrated.mesh.colors object (generated from rgb.calibrate) or mesh.colors object (generated from rgb.measure)}

\item{specimen.factors}{The covariate data that you want to combine with the calibrated data}

\item{use.perimeter.data}{If TRUE, we will return the color values for the perimeter landmarks as well as the interior landmarks.}

\item{write.csv}{A directory is to write a csv to, if desired}
}
\value{
The function will return a dataframe of specimen data (inherits the original column names) and landmark x & y values along with calibrated RGB values.
}
\description{
Combine metadata with calibrated colormesh data for analysis
}
