% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot.calibrated.mesh.colors}
\alias{plot.calibrated.mesh.colors}
\title{plotting individual specimens before AND after color sampling | select individual or 3 random side by side individuals}
\usage{
\method{plot}{calibrated.mesh.colors}(mesh.colors.object, individual = 1, style = "interior", ...)
}
\arguments{
\item{mesh.colors.object}{an object of class "calibrated.mesh.colors".}

\item{individual}{which individual from your landmark dataframe you'd like to plot}

\item{style}{options include "interior", "perimeter", "points", and "comparison".}

\item{...}{Additional plotting parameters to be passed to plot.default}
}
\value{
A plot of calibrated mesh colors
}
\description{
plotting individual specimens before AND after color sampling | select individual or 3 random side by side individuals
}
