% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_plotter.R
\name{point.overlap}
\alias{point.overlap}
\title{Diagnostic tool for looking at point sampling density}
\usage{
point.overlap(delaunay.map, px.radius, style = "points")
}
\arguments{
\item{delaunay.map}{delaunay triangulation object}

\item{px.radius}{The size of the circular neighborhood (in pixels) to sample color around each triangulated point.}

\item{style}{What kind of plot to show overlapping points with. Options are "points" and "triangulation".}
}
\value{
An index of points pairs that will have overlapping (and redundant) pixel information. Only for the interior points of a delaunay triangulation (not the perimeter)
}
\description{
Diagnostic tool for looking at point sampling density
}
\examples{

data(guppies)
delaunay.map <- tri.surf(raw.gup[,,1], point.map = c(1,8:17,2, 18:19,3,20:27,4, 28:42,5,43:52,6,53:54,7,55:62), 4, test.image)
point.overlap(delaunay.map, 2)
}
