% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb.measure.R
\name{rgb.measure}
\alias{rgb.measure}
\title{Color sampling from a set of pre-warped images}
\usage{
rgb.measure(
  imagedir,
  image.names,
  delaunay.map,
  px.radius = 2,
  linearize.color.space = F
)
}
\arguments{
\item{imagedir}{directory of images to measure. Only images with landmarks will be processed. The landmark file names are assumed to exactly match the image names.}

\item{image.names}{A vector of image names to look for in imagedir. These images should be unwarped or deformed to a common reference shape.}

\item{delaunay.map}{delaunay triangulation object}

\item{px.radius}{The size of the circular neighborhood (in pixels) to sample color around each triangulated point.}

\item{linearize.color.space}{should the sampled color data be transformed into linear color space}
}
\value{
The function will return $sampled.color-- an N_points x 3 (RGB) x N_observations array of sampled color values. A tri.surf.points class object will also be returned as $delaunay.
}
\description{
Color sampling from a set of pre-warped images
}
\examples{
#load landmarks and covariate data
guppy.lms <- tps2array(system.file("extdata", "original_lms.TPS", package = "Colormesh"))
specimen.factors <- read.csv(system.file("extdata", "specimen_factors.csv", package = "Colormesh"), header = F)
consensus <- tps2array(system.file("extdata", "consensus_LM_coords.TPS", package = "Colormesh"))
test.image <- load.image(paste0(path.package("Colormesh"),"/extdata/unwarped_images/GPHP_unw_001.jpg"))
delaunay.map <- tri.surf(consensus, point.map = c(1,8:17,2, 18:19,3,20:27,4, 28:42,5,43:52,6,53:54,7,55:62), 3, test.image)

rgb.test <- rgb.measure(imagedir = paste0(path.package("Colormesh"),"/extdata/unwarped_images/"), image.names = specimen.factors[,2], delaunay.map = delaunay.map, linearize.color.space = F)

plot(rgb.test, individual = 5)
plot(rgb.test, individual = 5, style = "comparison")

}
