% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tps.unwarp.R
\name{tps.unwarp}
\alias{tps.unwarp}
\title{Non-linear image registration using TPS deformation.}
\usage{
tps.unwarp(imagedir, landmarks, image.names, write.dir = NULL)
}
\arguments{
\item{imagedir}{directory of images to deform. Only images with landmarks will be processed. The landmark file names are assumed to exactly match the image names.}

\item{landmarks}{A landmark array with dimensions N_landmarks x 2 x N_observations. dimnames(landmarks)[[3]] should have the corresponding image filenames for each observation.}

\item{image.names}{A vector of image names to look for in imagedir. These images should be unwarped or deformed to a common reference shape.}

\item{write.dir}{Where to save warped images. Images will be named after the original image name (is that a bad idea because of overwriting? We will find out).}
}
\value{
warped images will be saved to the write.dir directory.
}
\description{
Non-linear image registration using TPS deformation.
}
\examples{
#load landmarks and covariate data
guppy.lms <- tps2array(system.file("extdata", "original_lms.TPS", package = "Colormesh"))
specimen.factors <- read.csv(system.file("extdata", "specimen_factors.csv", package = "Colormesh"), header = F)

#unwarp images
example.sample <- tps.unwarp(imagedir = paste0(path.package("Colormesh"),"/extdata/"), landmarks = guppy.lms, image.names = specimen.factors[,1], write.dir = tempdir())
}
