% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tps2array.R
\name{tps2array}
\alias{tps2array}
\title{read TPS files and convert data to an array}
\usage{
tps2array(data)
}
\arguments{
\item{data}{A .TPS file}
}
\value{
A matrix of the landmarks for each observation
}
\description{
read TPS files and convert data to an array
}
