% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tps2array.R
\name{tps2array}
\alias{tps2array}
\title{read TPS files and convert data to an array}
\usage{
tps2array(data)
}
\arguments{
\item{data}{A .TPS file}
}
\value{
An array of N_landmarks x 2 x N_observations
}
\description{
read TPS files and convert data to an array
}
\examples{
tps2array(system.file("extdata", "consensus_LM_coords.TPS", package = "Colormesh"))
}
